% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_onepulse.R
\name{gt_onepulse}
\alias{gt_onepulse}
\title{Convert a OnePulse Result to a GT Table}
\usage{
gt_onepulse(x, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{The result of \code{\link[=summarise_onepulse]{summarise_onepulse()}} or
\code{\link[=crosstab_onepulse]{crosstab_onepulse()}}.}

\item{title}{Optional table title.}

\item{subtitle}{Optional table subtitle.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Convert a OnePulse Result to a GT Table
}
